Remove the aggregation example files
- Use the AgentGenerator to generate the example agents code in the init-example command.
- Move the produce and init-example commands to the main package so that we can remove the example files entirely
- We don't need the aggregation topic name in the configuration anymore, it is now based on the source topic name and the topic_rename_format configuration
- Rename src_topic to source_topic_name to be consistent everywhere
