--- src/kafkaaggregator/example/__init__.py
@@ -1,5 +1,0 @@
-"""Aggregation example package.
-
-It shows an example on how to configure a Faust agent for data aggregation
-using the kafka-aggregator package.
-"""
